Add files via upload
The primary goal of this project was to determine the underlying reason for the high percentage of returned orders at Superstore. Through the use of a scatterplot that shows the connection between total sales and total returns. constructing a bar graph that shows the return rate by product type. Determine the return rate for each consumer to help you identify those who are more likely to return things. An area-specific map showing the return rate (state, city, etc.)  I look at the return rate for a specific time period (week, month, etc.) to see whether there is a seasonal influence on returned orders.  I was able to produce the necessary visualizations after creating bar graphs, scatterplots, and map charts. After that, I created a low fidelity mockup of my dashboard with a description of each chart and some previously created charts. while also creating filters that help make the content easier to understand for readers.